MOVE: cmd/kube-apiserver: move genericapiserver construction into controlplane package
