+++ pkg/controlplane/apiserver/.import-restrictions
@@ -1,0 +1,12 @@
+rules:
+  - selectorRegexp: k8s[.]io/kubernetes
+    allowedPrefixes:
+      - ''
+    forbiddenPrefixes:
+      # prevent pkg/ from depending on cmd/
+      #   note: pkg/kubemark overrides this
+      # - k8s.io/kubernetes/cmd # temporarily disabled until options are split and moved
+      # use sigs.k8s.io/yaml instead
+      - github.com/ghodss/yaml
+      # prevent kubernetes from opening sctp sockets (ref: https://github.com/kubernetes/kubernetes/pull/87926#discussion_r376642015)
+      - github.com/ishidawataru/sctp
